code cleanup/refactor, added issue update API endpoint
