+++ app/issues/[id]/EditIssueButton.tsx
@@ -1,4 +1,4 @@
-import { Pencil2Icon } from '@radix-ui/react-icons/dist/Pencil2Icon'
+import { Pencil2Icon } from '@radix-ui/react-icons'
 import { Button } from '@radix-ui/themes'
 import Link from 'next/link'
