+++ app/issues/page.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import delay from 'delay'
-import IssuesToolbar from './components/IssuesToolbar'
-import IssuesList from './components/IssuesList'
+import IssuesToolbar from './_components/IssuesToolbar'
+import IssuesList from './_components/IssuesList'
 
 const IssuesPage:React.FC = async () => {
     await delay(1000)
